Refactor: convert providers folder to ts
