--- src/providers/authProvider.jsx
@@ -1,23 +1,0 @@
-import { createContext, useContext, useEffect, useMemo, useState } from "react";
-
-const AuthContext = createContext();
-
-export const useAuthContext = () => {
-  return useContext(AuthContext);
-};
-
-export const AuthProvider = ({ children }) => {
-  const [user, setUser] = useState(JSON.parse(localStorage.getItem("user")));
-
-  useEffect(() => {
-    user
-      ? localStorage.setItem("user", JSON.stringify(user))
-      : localStorage.removeItem("user");
-  }, [user]);
-
-  const contextValue = useMemo(() => ({ user, setUser }), [user]);
-
-  return (
-    <AuthContext.Provider value={contextValue}>{children}</AuthContext.Provider>
-  );
-};
