+++ src/providers/authProvider.tsx
@@ -1,0 +1,47 @@
+import {
+  ReactNode,
+  Dispatch,
+  SetStateAction,
+  createContext,
+  useContext,
+  useEffect,
+  useMemo,
+  useState,
+} from "react";
+
+interface User {
+  id: string;
+  token: string;
+}
+
+interface AuthContextType {
+  user: User;
+  setUser: Dispatch<SetStateAction<User>>;
+}
+
+const AuthContext = createContext<AuthContextType | null>(null);
+
+export const useAuthContext = () => {
+  return useContext(AuthContext) as AuthContextType;
+};
+
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
+  const [user, setUser] = useState<User>(
+    JSON.parse(localStorage.getItem("user")!)
+  );
+
+  useEffect(() => {
+    user
+      ? localStorage.setItem("user", JSON.stringify(user))
+      : localStorage.removeItem("user");
+  }, [user]);
+
+  const contextValue: AuthContextType = useMemo(
+    () => ({ user, setUser }),
+    [user]
+  );
+
+  return (
+    <AuthContext.Provider value={contextValue}>{children}</AuthContext.Provider>
+  );
+};
